python/requirements.txt
# Mondrian Map - Python Data Processing Pipeline
# Install: python -m venv venv && venv\Scripts\activate && pip install -r requirements.txt

# Core data handling
pandas>=2.0.0
numpy>=1.24.0

# Enrichment analysis
gseapy>=1.0.1

# GoBERT embeddings
torch>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.16.0

# Dimensionality reduction & layout
umap-learn>=0.5.3
scikit-learn>=1.3.0

# Utilities
requests>=2.28.0
scipy>=1.10.0

# API server
flask>=3.0.0
flask-cors>=4.0.0
